packages/obs-title-updater/pyproject.toml
[project]
name = "obs-title-updater"
version = "0.1.0"
description = "OBS配信開始イベント連動のYouTube動画タイトル更新スクリプト"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.3",
    "pytz>=2024.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # line too long (handled by formatter)
    "E402", # module level import not at top of file (sys.path manipulation required)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
